hubble: move peer interface definitions to hubble/peer/types
[ upstream commit bee30fa7533bb45168e832ad06ca44266b49e4c8 ]

This is necessary to avoid import cycles when implementing fakes for the
peer interfaces in the `testutils` package.

Note that the `types` sub-package is commonly found throughout Cilium's
codebase so this change follows the convention.
